fix(config): consolidate next.config.js and remove duplicate next.config.ts
Unify all Next.js configuration into a single next.config.js to eliminate
the divergent duplicated config. Next 14 (installed via lockfile) only
loads next.config.js/next.config.mjs; the next.config.ts file was ignored
and had drifted (missing images.remotePatterns, bodySizeLimit, webpack
fallbacks, security headers).

Merged from next.config.ts (preserved):
- trailingSlash, generateEtags, reactStrictMode
- splitChunks vendor/common, alias
- i18n rewrites and redirects
- sitemap/robots content-type + no-store headers
- long-lived asset cache header

Merged from next.config.js (preserved):
- images.remotePatterns (all social platforms data sources)
- serverActions bodySizeLimit 10mb
- webpack resolve.fallback (fs/net/tls/crypto) on client
- security headers (CSP, HSTS, X-Frame-Options, etc.)

Verified: npm run build passes, npm run lint 0 errors.
